requirements.txt
Pillow
opencv-python
numpy
flask
flask-cors
pydub
gTTS
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai
schedule
pytz
psycopg2-binary
python-dotenv
opencv-python
numpy
flask
flask-cors
pydub
gTTS
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-generativeai
schedule
pytz
psycopg2-binary
